chore: move testcases to release group test file
